+++ src/app-config/app-config.ts
@@ -93,6 +93,7 @@ export const APP_CATCH_KEY_PREFIX = ''
  *   - appPiniaMenuStore: pinia menu store key
  *   - appPiniaSigningStore: pinia signing store key
  *   - appVersionProvider: 版本信息缓存 key
+ *   - appMenuTagOptions: 标签页菜单列表
  */
 export const APP_CATCH_KEY = {
   signing: 'signing',
@@ -106,6 +107,7 @@ export const APP_CATCH_KEY = {
   appVersionProvider: 'appVersionProvider',
   isAppLockScreen: 'isAppLockScreen',
   appGlobalSearchOptions: 'appGlobalSearchOptions',
+  appMenuTagOptions: 'appMenuTagOptions',
 } as const
 
 /**